HEXDEV-698: Refactoring - move AuthenticationHandler to AbstractHTTPD
